Refactor and rename code bases and make some stylistic changes
