+++ src/CodeSnippetMessage.ts
@@ -6,13 +6,11 @@ import { Dialog } from '@jupyterlab/apputils';
 
 import { Widget } from '@lumino/widgets';
 import { Message } from '@lumino/messaging';
-// import { PromiseDelegate } from '@lumino/coreutils';
-// import { ArrayExt } from '@lumino/algorithm';
 
 import checkSVGstr from '../style/icon/jupyter_checkmark.svg';
 
 /**
- * The class name for confirmation box
+ * The class name for message
  */
 // const CONFIRM_CLASS = 'jp-codeSnippet-confirm';
 const CONFIRM_CONTENT = 'jp-codeSnippet-Message-content';
@@ -26,18 +24,18 @@ const CODE_SNIPPET_CONFIRM_TEXT = 'jp-codeSnippet-confirm-text';
  *
  * @returns A promise that resolves with whether the dialog was accepted.
  */
-export function showMessage<T>(): Promise<Dialog.IResult<T>> {
-  const confirmMessage = new ConfirmMessage({
-    body: new MessageHandler(),
+export function showMessage<T>(type: string): Promise<Dialog.IResult<T>> {
+  const confirmMessage = new CodeSnippetMessage({
+    body: new MessageHandler(type),
     buttons: [],
   });
   return confirmMessage.launch();
 }
 
 /**
- * A widget used to show confirmation message.
+ * A widget used to show message.
  */
-export class ConfirmMessage<T> extends Dialog<any> {
+export class CodeSnippetMessage<T> extends Dialog<any> {
   constructor(options: Partial<Dialog.IOptions<T>> = {}) {
     super(options);
     this.children().next().addClass(CONFIRM_CONTENT);
@@ -45,7 +43,9 @@ export class ConfirmMessage<T> extends Dialog<any> {
 
   protected onAfterAttach(msg: Message): void {
     const node = this.node;
+
     node.addEventListener('click', this, true);
+    document.addEventListener('keydown', this, false);
   }
 
   /**
@@ -54,12 +54,35 @@ export class ConfirmMessage<T> extends Dialog<any> {
   protected onAfterDetach(msg: Message): void {
     const node = this.node;
     node.removeEventListener('click', this, true);
+    document.removeEventListener('keydown', this, false);
+  }
+
+  handleEvent(event: Event): void {
+    switch (event.type) {
+      case 'keydown':
+        this._evtKeydown(event as KeyboardEvent);
+        break;
+      case 'click':
+        this._evtClick(event as MouseEvent);
+        break;
+    }
+  }
+
+  protected _evtKeydown(event: KeyboardEvent): void {
+    // Check for escape key
+    switch (event.key) {
+      case 'Escape':
+        event.stopPropagation();
+        event.preventDefault();
+        this.reject();
+        break;
+    }
   }
 }
 
 class MessageHandler extends Widget {
-  constructor() {
-    super({ node: Private.createConfirmMessageNode() });
+  constructor(type: string) {
+    super({ node: Private.createMessageNode(type) });
     this.addClass(CONFIRM_BODY);
   }
 }
@@ -69,14 +92,20 @@ class MessageHandler extends Widget {
  */
 namespace Private {
   // create a confirm message when new snippet is created successfully
-  export function createConfirmMessageNode(): HTMLElement {
+  export function createMessageNode(type: string): HTMLElement {
     const body = document.createElement('div');
     body.innerHTML = checkSVGstr;
 
     const messageContainer = document.createElement('div');
     messageContainer.className = CODE_SNIPPET_CONFIRM_TEXT;
     const message = document.createElement('text');
-    message.textContent = 'Saved as Snippet!';
+    if (type === 'confirm') {
+      message.textContent = 'Saved as Snippet!';
+    } else if (type === 'copy') {
+      message.textContent = 'Saved to Clipboard!';
+    } else if (type === 'download') {
+      message.textContent = 'Downloaded the Snippet!';
+    }
     messageContainer.appendChild(message);
     body.append(messageContainer);
     return body;
